rewrite tests for BlockingQueue
this will allow the tests to have access to private variables
