+++ src/threads/blocking_queue.h
@@ -5,6 +5,11 @@
 #include <memory>
 #include <mutex>
 
+// If the unit tests are enabled, include the following header
+#ifdef TESTS_ENABLED
+  #include <gtest/gtest.h>
+#endif
+
 /**
  * This is a thread safe queue that blocks on a mutex for each operation
  **/
@@ -34,6 +39,17 @@ private:
   std::shared_ptr<Deque> internal_queue;
   // A mutex to lock the queue when performing operations
   std::shared_ptr<Mutex> mtx;
+
+  #ifdef TESTS_ENABLED
+    FRIEND_TEST(BlockingQueueTest, TEST_FRONT);
+    FRIEND_TEST(BlockingQueueTest, TEST_BACK);
+    FRIEND_TEST(BlockingQueueTest, TEST_SIZE);
+    FRIEND_TEST(BlockingQueueTest, TEST_EMPTY);
+    FRIEND_TEST(BlockingQueueTest, TEST_CLEAR);
+    FRIEND_TEST(BlockingQueueTest, TEST_PUSH);
+    FRIEND_TEST(BlockingQueueTest, TEST_PUSH_FRONT);
+    FRIEND_TEST(BlockingQueueTest, TEST_POP);
+  #endif
 };
 
 #endif // SRC_THREADS_BLOCKING_QUEUE_H
