+++ test/src/threads/blocking_queue.cc
@@ -1,4 +1,4 @@
-#include "blocking_queue_mock.h"
+#include "blocking_queue.h"
 #include "deque_mock.h"
 #include "mutex_mock.h"
 
@@ -6,29 +6,6 @@
 #include <gtest/gtest.h>
 
 using ::testing::Sequence;
-
-// Test Fixture for BlockingQueue class
-class BlockingQueueTest : public ::testing::Test
-{
-protected:
-  BlockingQueueTest()
-    : my_internal_queue{ new DequeMock<int>() },
-      mtx_mock{ new MutexMock() },
-      b_queue(my_internal_queue, mtx_mock)
-  {
-  }
-
-  virtual void SetUp() {}
-
-  void expect_locks(unsigned int num);
-
-  const int arg0 = 0;
-
-  std::shared_ptr<DequeMock<int>> my_internal_queue;
-  std::shared_ptr<MutexMock> mtx_mock;
-
-  BlockingQueueMock<int, DequeMock<int>, MutexMock> b_queue;
-};
 
 // Expect the mutex to be locked and unlocked `num` times
 // If lock/unlock is called in the wrong order this throws an error
